created user manager & model
